% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTrial_Maurer_Bretz_2013.R
\name{buildTrial_Maurer_Bretz_2013}
\alias{buildTrial_Maurer_Bretz_2013}
\title{Generate efficacy boundaries and then calculate trial performance using method from Maurer Bretz (2013)}
\usage{
buildTrial_Maurer_Bretz_2013(...)
}
\arguments{
\item{...}{passed to \code{\link{getEffBounds_Maurer_Bretz_2013}} and \code{\link{simTrial_Maurer_Bretz_2013}}.}
}
\description{
This function calls \code{\link{getEffBounds_Maurer_Bretz_2013}} and \code{\link{simTrial_Maurer_Bretz_2013}}.
}
\references{
Maurer, W. and F. Bretz (2013). Multiple testing in group sequential trials using graphical approaches. \emph{Statistics in Biopharmaceutical Research.}
}
