% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeTrial.r
\name{cleanArgList}
\alias{cleanArgList}
\title{Clean a list of arguments (for internal use)}
\usage{
cleanArgList(x, max_K, min_n_total = 10)
}
\arguments{
\item{x}{list of arguments}

\item{max_K}{maximum number of stages}

\item{min_n_total}{minimum per-stage sample size}
}
\description{
For internal use within scripts, and within \code{\link{optimizeTrial}}
}
