% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeTrial.r
\name{get_case_perf_obj}
\alias{get_case_perf_obj}
\title{Get trial performance for each case.}
\usage{
get_case_perf_obj(base_args, cases, objective_fun, trial_method,
  return_entries = c("performance"), ...)
}
\arguments{
\item{base_args}{the trial design and other inputs to functions determined by \code{trial_method}. This corresponds to the combined list of \code{args_list_init} and \code{args_list_fixed} from \code{\link{optimizeTrial}}.}

\item{cases}{analagous to the input for \code{\link{optimizeTrial}}.}

\item{objective_fun}{a function to evaluate the design, see \code{\link{optimizeTrial}}.}

\item{trial_method}{either 'MB' or 'cov', depending on the type of design being used. See \code{\link{optimizeTrial}}}

\item{return_entries}{a list of strings telling which elements of the returned value from \code{objective_fun} should be stored. Set to 'all' to return all elements from functions determined by \code{trial_method}.}

\item{...}{is passed to \code{objective_fun}}
}
\value{
A list with elements
\item{cases}{a modified version of \code{cases} in which each element has been appended with descriptive metrics from functions determined by \code{trial_method}.}
\item{obj}{The output from \code{objective_fun}}
}
\description{
This function is mostly for internal use within \code{\link{optimizeTrial}}. It first gets trial performance for each case. Once the characteristics of the trial in each case are simulated, the objective function can be calculated.
}
\details{
See argument definitions in \code{\link{optimizeTrial}} for more information.
}
