% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeTrial.r
\name{get_E_SS}
\alias{get_E_SS}
\title{Returns the expected sample size (mostly for internal use)}
\usage{
get_E_SS(cases)
}
\arguments{
\item{cases}{Describes the scenarios under which sample size should be calculated. Each case should include a \code{performance} element and a \code{weight} element.}
}
\value{
the expected sample size across cases.
}
\description{
Mostly for internal use. This function is called by \code{\link{get_case_perf_obj}}.
}
