% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeTrial.r
\name{get_perf_mat}
\alias{get_perf_mat}
\title{Get matrix with trial design performance in each supplied case}
\usage{
get_perf_mat(perf_list)
}
\arguments{
\item{perf_list}{output object from \code{\link{get_case_perf_obj}}}
}
\description{
This function parses the output of \code{\link{get_case_perf_obj}} into a more easily read form.
}
