% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeTrial.r
\name{getBoundsFromOptimSoln}
\alias{getBoundsFromOptimSoln}
\title{Get efficacy boundary from a solution found by \code{\link{optimizeTrial}}}
\usage{
getBoundsFromOptimSoln(soln, case, trial_method)
}
\arguments{
\item{soln}{the \code{soln} element in the returned list from \code{\link{optimizeTrial}}}

\item{case}{a scenario in which to calculate boundaries, in the same format as an element of the \code{cases} list argument for \code{\link{optimizeTrial}}}

\item{trial_method}{either 'cov' or 'MB'.}
}
\description{
This function helps to parse the output from \code{\link{optimizeTrial}} into a more easily viewed format.
}
