% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTrial_Maurer_Bretz_2013.R
\name{getEffBounds_Maurer_Bretz_2013}
\alias{getEffBounds_Maurer_Bretz_2013}
\title{Compute efficacy boundaries for an adaptive trial using method from Maurer Bretz (2013)}
\usage{
getEffBounds_Maurer_Bretz_2013(p1, r1, r2, var_s1_trt, var_s1_con, var_s2_trt,
  var_s2_con, time_limit, num_stages, n_total = NULL, n_per_stage, FWER,
  H01_eff_allocated = NULL, H02_eff_allocated = NULL,
  H0C_eff_allocated = NULL, FWER_allocation_matrix = NULL,
  delta_eff = NULL, H01_eff_total_allocated = NULL,
  H02_eff_total_allocated = NULL, H0C_eff_total_allocated = NULL, abseps,
  maxpts, errtol, graph_edge_12, graph_edge_2C, graph_edge_C1, ...)
}
\arguments{
\item{graph_edge_12}{The proportion of alpha to reallocate from H_01 to H_02 in the event that H_01 is rejected}

\item{graph_edge_2C}{The proportion of alpha to reallocate from H_02 to H_0C in the event that H_02 is rejected}

\item{graph_edge_C1}{The proportion of alpha to reallocate from H_0C to H_01 in the event that H_0C is rejected}
}
\description{
All arguments are analagous to \code{\link{getEffBounds}}, with the addition of three parameters: \code{graph_edge_12}, \code{graph_edge_2C}, and \code{graph_edge_C1}.
}
\references{
Maurer, W. and F. Bretz (2013). Multiple testing in group sequential trials using graphical approaches. \emph{Statistics in Biopharmaceutical Research.}
}
