% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleInputsMISTIE.R
\name{getExampleInputsMISTIE}
\alias{getExampleInputsMISTIE}
\title{Example inputs for \code{\link{optimizeTrial}}}
\usage{
getExampleInputsMISTIE(max_K = 4, trial_method = "cov",
  bound_style = "unstructured", which_obj = "ss", iter = 50,
  showiter = FALSE)
}
\arguments{
\item{max_K}{maximum number of stages}

\item{trial_method}{either 'cov' or 'MB'}

\item{bound_style}{either 'unstructured' or 'structured'}

\item{which_obj}{either 'ss' or 'dur' to minimize expected sample size or duration respectively}

\item{iter}{number of search iterations}

\item{showiter}{passed to \code{\link{min_n_feasible}}}
}
\description{
Example inputs for \code{\link{optimizeTrial}}
}
\seealso{
\code{\link{optimizeTrial}}
}
