% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeTrial.r
\name{min_E_SS_power_constraints}
\alias{min_E_SS_power_constraints}
\alias{min_E_dur_power_constraints}
\title{Sample objective function to minimize}
\usage{
min_E_SS_power_constraints(...)

min_E_dur_power_constraints(...)
}
\arguments{
\item{...}{for internal use.}
}
\value{
a list containing
\item{base}{the expected sample size, or expected duration}
\item{penalized}{the value of \code{base} plus a penalty if any of the power constraints are violated}
\item{power_diffs}{a matrix showing the difference between achieved power and required power for each case, and each constraint.}
}
\description{
\code{\link{optimizeTrial}} requires users to input an objective function to evaluate the trial design in each case. \code{min_E_SS_power_constraints} is a pre-set option for an objective function based on expected sample size. \code{min_E_dur_power_constraints} is a pre-set option for an objective function based on expected duration. Both functions' output a list including a penalized objective function value. Any custom objective function supplied to  \code{\link{optimizeTrial}} must be able to run in the context of the \code{\link{get_case_perf_obj}} function, and give output in the same format as that of the functions below.
}
