% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sann_stochastic.R
\name{sann_random}
\alias{sann_random}
\title{A simulated annealing variant that re-evaluates the reference point at each comparison}
\usage{
sann_random(fn, par, control = list())
}
\arguments{
\item{fn}{function to minimize}

\item{par}{initial parameters}

\item{control}{analogous to \code{\link{optim}}}
}
\value{
optimized value, found at minimum of all evaluated solutions.
}
\description{
A simulated annealing variant that re-evaluates the reference point at each comparison
}
\examples{
\dontrun{
fn<-function(x){
    abs(x) + rnorm(n=1,mean=0,sd=.3) 
}
par = 2
control <- list()
control$parscale <- 2
control$maxit <- 3000
control$temp<-10
sann_random(fn, par, control)
}
}
