% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTrial.R, R/buildTrial_Maurer_Bretz_2013.R
\name{simTrial}
\alias{simTrial}
\alias{simTrial_Maurer_Bretz_2013}
\title{Simulate a trial to compute power, expected sample size, and expected duration}
\usage{
simTrial(p1, r1, r2, mean_s1_trt = NULL, mean_s1_con = NULL,
  mean_s2_trt = NULL, mean_s2_con = NULL, var_s1_trt = NULL,
  var_s1_con = NULL, var_s2_trt = NULL, var_s2_con = NULL, iter,
  time_limit = 90, num_stages, n_total = NULL, n_per_stage,
  all_efficacy_boundaries = NULL, H01_efficacy_boundaries = NULL,
  H02_efficacy_boundaries = NULL, H0C_efficacy_boundaries = NULL,
  H01_futility_boundaries = NULL, H02_futility_boundaries = NULL,
  H0C_futility_boundaries = NULL, delta_futility = NULL,
  intercepts_futility = NULL, H01_futility_boundary_const = NULL,
  H02_futility_boundary_const = NULL, H0C_futility_boundary_const = NULL,
  enrollment_rate_combined, delay, ...)

simTrial_Maurer_Bretz_2013(p1, r1, r2, mean_s1_trt, mean_s1_con, mean_s2_trt,
  mean_s2_con, var_s1_trt, var_s1_con, var_s2_trt, var_s2_con, iter, time_limit,
  num_stages, n_total = NULL, n_per_stage, all_efficacy_boundaries = NULL,
  H01_futility_boundaries, H02_futility_boundaries, H0C_futility_boundaries,
  delta_futility = NULL, intercepts_futility = NULL,
  H01_futility_boundary_const = NULL, H02_futility_boundary_const = NULL,
  H0C_futility_boundary_const = NULL, enrollment_rate_combined, delay, ...)
}
\arguments{
\item{p1}{Proportion of population in subpopulation 1.}

\item{r1}{probability of being randomized to treatment in subpopulation 1}

\item{r2}{probability of being randomized to treatment in subpopulation 2}

\item{mean_s1_trt}{mean of the outcome under treatment in subpopluation 1.}

\item{mean_s1_con}{mean of the outcome under control in subpopluation 1.}

\item{mean_s2_trt}{mean of the outcome under treatment in subpopluation 2.}

\item{mean_s2_con}{mean of the outcome under control in subpopluation 2.}

\item{var_s1_trt}{variance of the outcome under treatment in subpopluation 1.}

\item{var_s1_con}{variance of the outcome under control in subpopluation 1.}

\item{var_s2_trt}{variance of the outcome under treatment in subpopluation 2.}

\item{var_s2_con}{variance of the outcome under control in subpopluation 2.}

\item{iter}{The number of simulated trials used to
estimate the power, expected sample size, and expected trial duration.}

\item{time_limit}{time limit for calculations.}

\item{num_stages}{Total number of stages
used in each design  (\eqn{K}).  The maximum allowable number of stages is 20.}

\item{n_total}{the total, maximum number of patients to recruit by the end of the study. If entered, n_per_stage will be scaled to have this sum.}

\item{n_per_stage}{a vector with length equal to \code{num_stages}, telling the number of patients to enroll in each stage.}

\item{all_efficacy_boundaries}{a list of efficacy boundaries matching the output of \code{\link{getEffBounds}} for \code{simTrial}, or of \code{\link{getEffBounds_Maurer_Bretz_2013}} for \code{simTrial_Maurer_Bretz_2013}.}

\item{H01_efficacy_boundaries}{rather than setting \code{all_efficacy_boundaries}, the user can enter vectors for \code{H01_efficacy_boundaries}, \code{H02_efficacy_boundaries}, and \code{H0C_efficacy_boundaries}.}

\item{H02_efficacy_boundaries}{see \code{H01_efficacy_boundaries}}

\item{H0C_efficacy_boundaries}{see \code{H01_efficacy_boundaries}}

\item{H01_futility_boundaries}{a vector of futility boundaries for the hypothesis \eqn{H01}.}

\item{H02_futility_boundaries}{a vector of futility boundaries for the hypothesis \eqn{H02}.}

\item{H0C_futility_boundaries}{Not currently used in the algorithm, but may be added in the future.}

\item{delta_futility}{rather than setting the specific futility boundaries, parametric boundaries can be calculated. See \code{\link{getSmoothBounds}}.}

\item{intercepts_futility}{for use in \code{\link{getSmoothBounds}}}

\item{H01_futility_boundary_const}{for use in \code{\link{getSmoothBounds}}}

\item{H02_futility_boundary_const}{for use in \code{\link{getSmoothBounds}}}

\item{H0C_futility_boundary_const}{for use in \code{\link{getSmoothBounds}}}

\item{enrollment_rate_combined}{The assumed
enrollment rate per year for the combined population.  This impacts the
expected duration of each trial design. Active enrollments from
the two subpopulations are assumed to be independent.  The enrollment rates
for subpopulations 1 and 2 are assumed proportional, based on \code{p_1}.
This implies that each stage of the adaptive design up to and including stage \code{k*} takes the same amount of time to complete, regardless of whether or not enrollment stops for subpopulation 2.  Each stage after \code{k*} will also take the same amount of time to complete.}

\item{delay}{delay time from participant enrollment to observation    of his/her outcome (in years)}

\item{...}{needed so that function ignores unused arguments when called by \code{\link{buildTrial}}}
}
\description{
\code{simTrial} simulates an adaptive enrichment trial design, using the method from Rosenblum et al. (2016), in order to compute the following
performance criteria: power, expected sample size, and expected duration. \code{simTrial_Maurer_Bretz_2013} follows an analogous procedure using the method from Maurer & Bretz (2013). \cr \cr Let \eqn{H01}, \eqn{H02} and \eqn{H0C} respectively denote the null hypotheses that there is no treatment effect in subpopulation 1, subpopulation 2 and the combined population.
}
\details{
For \code{simTrial}, first, cumulative Z-statistics are constructed for each stage and population.
Next, the enrollment modification rule and multiple testing procedure are applied
at each stage, which determines when accrual is stopped for each subpopulation
and when (if at all) each population's null hypothesis is rejected. 
\cr \cr If efficacy boundaries have not yet been computed, the user should consider using \code{\link{buildTrial}} or \code{\link{buildTrial_Maurer_Bretz_2013}}, which automatically complete this precursor step.
}
\references{
Maurer, W. and Bretz, F. (2013). Multiple testing in group sequential trials using graphical approaches. \emph{Statistics in Biopharmaceutical Research.}
\cr \cr
Rosenblum, M., Qian, T., Du, Y., and Qiu, H., Fisher, A. (2016). Multiple Testing Procedures for Adaptive Enrichment Designs: Combining Group Sequential and Reallocation Approaches. \emph{Biostatistics}. 17(4), 650-662. https://goo.gl/c8GlcH
}
