% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTrial_Maurer_Bretz_2013.R
\name{simTrial_Maurer_Bretz_2013}
\alias{simTrial_Maurer_Bretz_2013}
\title{Simulate a trial to compute its power, expected sample size, and expected duration using method from Maurer Bretz (2013)}
\usage{
simTrial_Maurer_Bretz_2013(p1, r1, r2, mean_s1_trt, mean_s1_con, mean_s2_trt,
  mean_s2_con, var_s1_trt, var_s1_con, var_s2_trt, var_s2_con, iter, time_limit,
  num_stages, n_total = NULL, n_per_stage, all_efficacy_boundaries = NULL,
  H01_futility_boundaries, H02_futility_boundaries, H0C_futility_boundaries,
  delta_futility = NULL, intercepts_futility = NULL,
  H01_futility_boundary_const = NULL, H02_futility_boundary_const = NULL,
  H0C_futility_boundary_const = NULL, enrollment_rate_combined, delay, ...)
}
\arguments{
\item{...}{passed to \code{\link{getEffBounds_Maurer_Bretz_2013}}}
}
\description{
All arguments are analagous to \code{\link{simTrial}}. However, additional parameters are required by \code{\link{getEffBounds_Maurer_Bretz_2013}}.
}
\references{
Maurer, W. and F. Bretz (2013). Multiple testing in group sequential trials using graphical approaches. \emph{Statistics in Biopharmaceutical Research.}
}
