% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTrial.R
\name{simTrial}
\alias{simTrial}
\title{Simulate a trial to compute its power, expected sample size, and expected duration}
\usage{
simTrial(p1, r1, r2, mean_s1_trt = NULL, mean_s1_con = NULL,
  mean_s2_trt = NULL, mean_s2_con = NULL, var_s1_trt = NULL,
  var_s1_con = NULL, var_s2_trt = NULL, var_s2_con = NULL, iter,
  time_limit = 90, num_stages, n_total = NULL, n_per_stage,
  all_efficacy_boundaries = NULL, H01_efficacy_boundaries = NULL,
  H02_efficacy_boundaries = NULL, H0C_efficacy_boundaries = NULL,
  H01_futility_boundaries = NULL, H02_futility_boundaries = NULL,
  H0C_futility_boundaries = NULL, delta_futility = NULL,
  intercepts_futility = NULL, H01_futility_boundary_const = NULL,
  H02_futility_boundary_const = NULL, H0C_futility_boundary_const = NULL,
  enrollment_rate_combined, delay, ...)
}
\arguments{
\item{p1}{Proportion of population in subpopulation 1.}

\item{r1}{probability of being randomized to treatment in subpopulation 1}

\item{r2}{probability of being randomized to treatment in subpopulation 2}

\item{mean_s1_trt}{mean of the outcome under treament in subpopluation 1.}

\item{mean_s1_con}{mean of the outcome under control in subpopluation 1.}

\item{mean_s2_trt}{mean of the outcome under treament in subpopluation 2.}

\item{mean_s2_con}{mean of the outcome under control in subpopluation 2.}

\item{var_s1_trt}{variance of the outcome under treament in subpopluation 1.}

\item{var_s1_con}{variance of the outcome under control in subpopluation 1.}

\item{var_s2_trt}{variance of the outcome under treament in subpopluation 2.}

\item{var_s2_con}{variance of the outcome under control in subpopluation 2.}

\item{iter}{The number of simulated trials used to
estimate the power, expected sample size, and expected trial duration.}

\item{time_limit}{time limit for calculations.}

\item{num_stages}{Total number of stages
used in each design  (\eqn{K}).  The maximum allowable number of stages is 20.}

\item{n_total}{the total, maximum number of patients to recruit by the end of the study. If entered, n_per_stage will be scaled to have this sum.}

\item{n_per_stage}{a vector with length equal to \code{num_stages}, telling the number of patients to enroll in each stage.}

\item{all_efficacy_boundaries}{a list of efficacy boundaries matching the output of \code{\link{getEffBounds}}}

\item{H01_efficacy_boundaries}{rather than setting \code{all_efficacy_boundaries}, the user can enter vectors for \code{H01_efficacy_boundaries}, \code{H02_efficacy_boundaries}, and \code{H0C_efficacy_boundaries}.}

\item{H02_efficacy_boundaries}{see \code{H01_efficacy_boundaries}}

\item{H0C_efficacy_boundaries}{see \code{H01_efficacy_boundaries}}

\item{H01_futility_boundaries}{a vector of futility boundaries for the hypothesis \eqn{H01}.}

\item{H02_futility_boundaries}{a vector of futility boundaries for the hypothesis \eqn{H02}.}

\item{H0C_futility_boundaries}{Not currently used in the algorithm, but may be added in the future.}

\item{delta_futility}{rather than setting the specific futility boundaries, parametric boundaries can be calculated. See \code{\link{getSmoothBounds}}.}

\item{H01_futility_boundary_const}{for use in \code{\link{getSmoothBounds}}}

\item{H02_futility_boundary_const}{for use in \code{\link{getSmoothBounds}}}

\item{H0C_futility_boundary_const}{for use in \code{\link{getSmoothBounds}}}

\item{enrollment_rate_combined}{The assumed
enrollment rate per year for the combined population.  This impacts the
expected duration of each trial design. Active enrollments from
the two subpopulations are assumed to be independent.  The enrollment rates
for subpopulations 1 and 2 are assumed proportional, based on \code{p_1}.
This implies that each stage of the adaptive design up to and including stage \code{k*} takes the same amount of time to complete, regardless of whether or not enrollment stops for subpopulation 2.  Each stage after \code{k*} will also take the same amount of time to complete.}

\item{delay}{delay time from participant enrollment to observation    of his/her outcome (in years)}

\item{...}{needed so that function ignores unused arguments when called by \code{\link{buildTrial}}}
}
\description{
If efficacy boundaries have not yet been computed, the user should consider using \code{\link{buildTrial}}, which automatically completes this precursor step.
 
Let \eqn{H01}, \eqn{H02} and \eqn{H0C} respectively denote the null hypotheses that there is no treatment effect in subpopulation 1, subpopulation 2 and the combined population.
}
\details{
This function is meant to be applied when there is prior
evidence that a treatment might work better in a one subpopulation
than in another. In this context, a trial with an adaptive enrollment
criteria would determine whether or not to continue enrolling patients
from each subpopulation based on interim analyses of whether each
subpopulation is benefiting. In order for the type I error and the
power of the trial to be calculable, the decision rules for changing
enrollment must be set before the trial starts. This function simulates trials with decision rules composed of efficacy boundaries and futility boundaries for \eqn{H01}, \eqn{H02}, and \eqn{H0C}, and reports the performance of the trial in
terms of power, expected sample size, and expected trial duration.
}
